requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
pyotp==2.9.0
qrcode==7.4.2
Pillow
werkzeug
